+++ Assets/_Scripts/Bots/BotController.cs
@@ -1,5 +1,6 @@
 using System;
 using UnityEngine;
+using UnityEngine.AI;
 
 namespace _Scripts.Bots
 {
@@ -22,9 +23,11 @@ namespace _Scripts.Bots
 
         [SerializeField] private float searchRadius = 10;
         [SerializeField] private float chaseRadius = 5;
-        [SerializeField] private float attackRadius = 1;
+        [SerializeField] public float attackRadius = 1;
         [SerializeField] private LayerMask mask;
 
+        public Transform target;
+        
         public int isIdlingHash;
         public int isWalkingHash;
         public int isRunningHash;
@@ -37,13 +40,15 @@ namespace _Scripts.Bots
             _movement = new BotMovement(gameObject);
             _animator = GetComponent<Animator>();
 
+            target = PlayerManager.instance.player.transform;
+
             isIdlingHash = Animator.StringToHash("idle");
             isWalkingHash = Animator.StringToHash("walking");
             isRunningHash = Animator.StringToHash("running");
             isAttackingHash = Animator.StringToHash("attacking");
             isDyingHash = Animator.StringToHash("dying");
 
-            _stats = new BotStats(2, 2);
+            _stats = new BotStats(50, 50, 5, 3);
         }
 
         // Update is called once per frame
@@ -52,6 +57,8 @@ namespace _Scripts.Bots
             UpdateBotMode();
 
             _animator.SetBool(isAttackingHash, false);
+
+            var distance = Vector3.Distance(target.position, transform.position);
             
             switch (_mode)
             {
@@ -65,36 +72,40 @@ namespace _Scripts.Bots
                     _animator.SetTrigger(isWalkingHash);
                     // _animator.SetBool(isAttackingHash, false);
                     // move to a random position which is near the enemy
+                    _movement.Move(target.position);            // demo
                     break;
                 case BotMode.ChaseEnemy:
                     Debug.Log("Bot is in chase mode");
                     _animator.SetTrigger(isRunningHash);
                     // _animator.SetBool(isAttackingHash, false);
                     // move directly to enemy position
+                    _movement.Move(target.position);
                     break;
                 case BotMode.AttackEnemy:
                     Debug.Log("Bot is in attack mode");
                     _animator.SetBool(isAttackingHash, true);
                     // stop movement and attack enemy
+                    // Attack the target
+                    FaceTarget();
                     break;
                 case BotMode.Die:
                     Debug.Log("Bot is in die mode");
                     _animator.SetTrigger(isDyingHash);
+                    Die();
                     break;
             }
 
-            if (Input.GetKeyDown(KeyCode.K))
-            {
-                TakeDamage(20);
-                Debug.Log(_stats.CurrentHealth);
-            }
         }
 
         private void UpdateBotMode()
         {
-            _mode = BotMode.Default;
-
             var pos = transform.position;
+            
+            // kein Gegner in range && wenn ich mein Ziel erreicht habe
+            if (!Physics.CheckSphere(pos, searchRadius, mask) && _movement.destinationReached)
+            {
+                _mode = BotMode.Default;
+            }
 
             // Check if enemy in search range
             if (Physics.CheckSphere(pos, searchRadius, mask))
@@ -149,9 +160,22 @@ namespace _Scripts.Bots
             Gizmos.DrawWireSphere(pos, attackRadius);
         }
 
-        private void TakeDamage(int dmg)
+        public void TakeDamage(int dmg)
         {
             _stats.TakeDamage(2);
+        }
+        
+        void FaceTarget()
+        {
+            Vector3 direction = (target.position - transform.position).normalized;
+            Quaternion lookRotation = Quaternion.LookRotation(new Vector3(direction.x, 0, direction.z));
+            transform.rotation = Quaternion.Slerp(transform.rotation, lookRotation, Time.deltaTime * 5f);
+        }
+
+        void Die()
+        {
+            // if()
+            Destroy(gameObject);
         }
     }
 }