+++ Assets/_Scripts/Player/PlayerController.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using _Scripts.Player;
+using UnityEngine;
+
+public class PlayerController : MonoBehaviour
+{
+    private PlayerStats _stats;
+
+    private void Start()
+    {
+        _stats = new PlayerStats(10, 10, 10, 3);
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}