+++ Assets/_Scripts/Player/Weapon.cs
@@ -1,0 +1,37 @@
+using _Scripts.Bots;
+using UnityEngine;
+
+namespace _Scripts.Player
+{
+    public class Weapon : MonoBehaviour
+    {
+        public Camera fpsCam;
+
+        public int damage = 10;
+        public float range = 100f;
+
+        // Update is called once per frame
+        void Update()
+        {
+            if (Input.GetButtonDown("Fire1"))
+            {
+                Shoot();
+            }
+        }
+
+        void Shoot()
+        {
+            RaycastHit hit;
+            if (Physics.Raycast(fpsCam.transform.position, fpsCam.transform.forward, out hit, range))
+            {
+                Debug.Log(hit.transform.name);
+
+                var enemy = hit.transform.GetComponent<BotController>();
+                if (enemy != null)
+                {
+                    enemy.TakeDamage(damage);
+                }
+            }
+        }
+    }
+}