+++ Assets/_Scripts/PlayerManager.cs
@@ -1,0 +1,20 @@
+using UnityEngine;
+
+namespace _Scripts
+{
+    public class PlayerManager : MonoBehaviour
+    {
+        #region Singleton
+
+        public static PlayerManager instance;
+
+        private void Awake()
+        {
+            instance = this;
+        }
+
+        #endregion
+
+        public GameObject player;
+    }
+}